Settings + noexcept + reduced compile-time (#82)
* settings + noexcept + ODR fix + papi_rate_tuple + kokkos

- removed some extranous interface libraries
- reworked tim::settings to be more dynamic
- added noexcept specifiers in many places
- fixed ODR issue with manager
- using cuda implies using nvtx
- rusage metrics value reports last diff
- child timing functions
- miscellaneous optimizations
- auto timer is simpler
- ert and storage are now a global lib
- support for embedding interpreter
- re-worked kokkos + more kokkos configs
- timem supports intermediate output
- TIMEMORY_DISABLE_CUDA_HALF -> TIMEMORY_USE_CUDA_HALF

* Update timemory-config.cmake.in

- tweak to deprecated interface library removal

* cmake, warning, and obsolete function fixes

* Fix settings display in timemory-avail

* fixes for cupti_tests and warnings about double-promotion

* Template instantiations of the roofline components

* Update ert/counter.hpp

* WIN32_LEAN_AND_MEAN, Signals, macros/os.hpp

- included macros/os.hpp where necessary
- renamed sys_signals python enum to Signals

* Update libpytimemory-settings.cpp

- fixes import error caused by overloading instance and static method not being supported

* Update derived_tests.cpp

* Update ert.hpp

* Fixes for Windows builds

* Update timemory.hpp

- use the top-level headers

* Update derived_tests.cpp

- enable signal detection

* Python build fixes for settings on Windows

- hold instance of settings shared_ptr in manager
- remove testing.hpp header from timemory/utility.hpp
- revert exec_data move ctor/assign to non-noexcept version

* Update .appveyor.yml

- BUILD_STATIC variant that is opposite of EXAMPLES variant

* Formatting and header include updates

- improved some header includes to reduce compile-time

* insert_node -> push_node + new operations + start/stop mods

- renamed insert_node to push_node
- created operation::cleanup, operation::extra_serialization
- removed base class start/stop/cleanup/extra_serialization implementations
- support extra_serialization member function with one argument

* Update declaration.hpp

* Windows fixes

* Update ert/data.hpp

- explicitly define exec_data move ctor and assignment due to strange undefined references with some compilers

* Updates storage/definition.hpp

* Fix to Windows builds

- removed extern template for finalize::flamegraph

* Update ert/data.hpp

* operation::merge + basic_tree + node::tree

- improved basic_tree layout via node::tree + merging
- fixes for test_throttle
- fixes for papi_tuple printing
- exclude dummy nodes in basic_tree
- removed debug output from timemory-avail
-

* header updates + ert_timer (minimal version of wall_clock)

- ert_timer reduces compile time

* CMake updates + kokkosp in main library

- use CMAKE_INTERPROCEDURAL_OPTIMIZATION for LTO
- support unity build
- CMAKE_VISIBILITY_INLINES_HIDDEN

* CMake fixes and init/fini stability improvements

- update travis
- replaced CMAKE_{SOURCE,BINARY} with PROJECT where relevant
- unity build fixes
- stability fixes with init/fini
- reduced default signals
- updated caliper, pybind11, and gotcha submodules
- TIMEMORY_UNITY_BUILD

* Update gotcha

* gotcha submodule + sockets

- socket tests

* Moved utility/serializer.hpp to cereal subfolder

- other minor cmake changes

* Windows winsock header fixes

- https://www.zachburlingame.com/2011/05/resolving-redefinition-errors-betwen-ws2def-h-and-winsock-h/

* Update .travis.yml

* Fix for CMP0077 in ex-ert

* Tweaks to cereal includes

* moved timemory/cereal to timemory/tpls/cereal

- windows had issue with "cereal/cereal" header included in cereal header

* Fixed issue with cereal includes

* test fixes for throttle fixes and component_bundle_tests

* Update .travis.yml

* Disable ERT example when building with LTO in CI

* Update ex_caliper.py

* Update CMakeLists.txt

* Minor changes to cmake examples

* Update .travis.yml

* Disable CMAKE_VISIBILITY_INLINES_HIDDEN for clang < 6.0

+++ source/tests/ompt_handle_tests.cpp
@@ -43,9 +43,10 @@ extern "C"
     extern void     timemory_deregister_ompt();
 }
 
-static const double pi_epsilon = std::numeric_limits<float>::epsilon();
-static const long   nfib       = 39;
-static const long   noff       = 4;
+static const double pi_epsilon =
+    static_cast<double>(std::numeric_limits<float>::epsilon());
+static const long nfib = 39;
+static const long noff = 4;
 
 //--------------------------------------------------------------------------------------//
