+++ source/tests/threading_tests.cpp
@@ -41,7 +41,8 @@ using namespace tim::component;
 using tuple_t = tim::auto_tuple_t<wall_clock, cpu_clock, cpu_util, thread_cpu_clock>;
 
 #if defined(_OPENMP)
-static const double pi_epsilon = std::numeric_limits<float>::epsilon();
+static const double pi_epsilon =
+    static_cast<double>(std::numeric_limits<float>::epsilon());
 #endif
 
 static const long nfib = 39;
